Adjust height of form and start styling button
Change the height of form to create more
space below the password fields. Change
width of input so it is not too long.

Relocate the declarations and styles
of the `button-div`to below, close to that
of `button` so both declarations
can be close to each other. Start styling
the button

On the index.html, add the paragraph below the
button