Create a container for the logo & its text
Add the logo image and its paragraph inside
the new container and start styling. Give it
a flex display, margin top (could have used
positioning here), a semi-transparent
background color, a fixed height and a 100%
width (to span 100% of its container).

Give the logo image fixed width and height
so it can be a reasonable size to fit in with
the paragraph text. Also style the paragraph
with the custom font of `norse-bold`.