Import and use the forum-regular font
Download the `Forum-Regular` from Google
font and save it in the fonts folder
(Downloading the font is one
of the ways to make Google fonts GDPR.
Importing it directly from Google font
library violates GDPR by collecting user's
IP addresses). Import it as a custom font
in the CSS and use it to style the all the h3
in the `rhs-container`.

Change the input's padding size to create
more space inside the input box. Also
give it a border style and a border radius