Style label using previous sibling combinator
On line 151, use the previous sibling combinator
to style the labels so that they become red if
the input below it is invalid.

On line 161 For all all labels of required inputs
not to be red on page load (all required inputs
are invalid when empty), style the label in such
a way that if the input is invalid and the placeholder
is shown (means the user has not started typing in
the input field), then the corresponding label would
keep its black color.