Style labels and inputs for `:valid` case
Style the labels and inputs in such a way
that when the input is valid, the border
is green and the background color is
transparent green.

Also when the input has focus and is valid,
the border stays green but the background
becomes transparent so that the user can
type without the green background.

On line 160, use the previous sibling
combinator to style the label that has a valid
input with a green color.

Add the `:placeholder-shown` class for the
`:valid` case beside that of the `:invalid`
case so that the labels of all non-required
inputs (non-required inputs are valid when
when empty) would not be green on page-load.