Add Confirm Password feature
Add a span below the first password
field in the HTML that would tell
users if their passwords match. Also
add a onkeyup event listener on
the password fields and the
passwordConfirm() function that
would be called onkeyup

On the javascript.js create the
passwordConfirm() function. If any of the
fields are empty, the span would be empty
so that a message would only output if
both fields are populated. The function
then outputs different messages if the
values of both fields are the same or not.