Set custom validation message.
Tweak the passwordConfirm() by using the
setCustomValidity() method
of the Constraint Validation API to set
custom validity messages when passwords
do not match. Also leave the
custom validity message blank if
the passwords match.
(If I have a custom message, the fields
are invalid but if I leave it blank "",
it becomes valid)

This makes the password fields invalid if
the both fields do not match.

Also change the class of the
.pass-match-info span to green or red if passwords
match or they don't, respectively.

On the CSS, create the .green and .red class
declaration and give them appropriate styles.