Adjust widths for responsiveness
Change the sidebar-container width to 37%
and the rhs-container width to 63% so it can
look like the Odin's sample.

Adjust the width of the inputs so that the
widths of the inputs can shrink as the width
of the form shrinks.

Style the hover state of the Login link

On the HTML change the alternate text of the
logo to some text that better describes it.