Add RegEx patterns to some fields
Add RexEx patterns to the First Name
and Last Name fields in the HTML so that
users cannot enter any number in those
fields. Also add a RegEx pattern to the
phone number field to match an US number
format.

On the HTML, Add a paragraph below
the first form h3 to indicate to users
which fields are required. Also mark
the required fields by the indicator
stated in the paragraph.

Adjust the padding (top) of the form
so that this paragraph can be styled
to have some space in between it, the
h3 text above it and the labels below
it.