Add the image-credits paragraph
In the HTML, add the paragraph,
the anchor tags and the links.

Give the paragraph's container a
relative position so that the
paragraph can positioned relative
to its container by the absolute
position. Also give top and left
properties to the paragraph and
a subtle hover effect to the
anchor tags inside the paragraph.

Tweak the selectors on lines 54,
65 and 70 to better select the
intended elements.