Style the labels and button
Change the font properties of
the labels and the paragraph
that is a child of the button-div.
Also change the color of the
labels (valid or invalid when
placeholder is shown).

Give the button a subtle
box-shadow and remove the box-shadow
on hover.

Adjust the left property of the
image-credits so that it will look
more like in the center when the page
is shrunk.