Add more validation rules to inputs
Use the minlength and maxlength
attributes to add more validation
rules to the first name and last name
fields and the password fields

The two name fields can not be less than
3 and greater than 12 chars. The password
fields cannot be less than 6 and greater
than 12 chars.

Refactor the passwordConfirm() function
in the script to set custom validity
message and change the pass-match-info
span to red with informative text if less
than 6 characters are typed into both of
the password fields.

Moved this length logic above the
password-matching logic in the if else
block because it seem more logical.
(The reverse also works.)

The little bug is that once the
pass-match-info span is displayed
(for valid or invalid), it does not
leave even when the password fields
are cleared out.

Adjust the left property on the
image-credits so it looks more
like in the center. Also adjust the
padding(top) of the .introduction
so that the tip of the button can
be visible before scrolling.

Also the inputmode="numeric" in the
phone number input is to change the
keyboard to numbers for users with
dynamic keyboards.